feat: pricing plugin seperated from governance
